+++ packages/cheeket-koa/lib/dependency.ts
@@ -1,9 +1,8 @@
 import Application, { DefaultContext, DefaultState } from "koa";
-import { Container, interfaces } from "cheeket";
+import { interfaces } from "cheeket";
 
 import ContainerContext from "./container-context";
 import * as Token from "./token";
-import Initializer from "./initializer";
 
 function container<
   StateT = DefaultState,
@@ -11,16 +10,13 @@ function container<
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   ResponseBodyT = any
 >(
-  initializer: Initializer,
+  rootContainer: interfaces.Container,
   options?: interfaces.ContainerConstructorOptions
 ): Application.Middleware<
   StateT,
   ContextT & Partial<ContainerContext>,
   ResponseBodyT
 > {
-  const rootContainer = new Container(options);
-  initializer.initRootContainer(rootContainer);
-
   return async (ctx, next) => {
     const contextContainer = rootContainer.createChildContainer(options);
 
@@ -34,8 +30,6 @@ function container<
     contextContainer.bind(Token.Cookies, () => ctx.cookies);
     contextContainer.bind(Token.Accepts, () => ctx.accept);
     contextContainer.bind(Token.Respond, () => ctx.respond);
-
-    initializer.initContextContainer(contextContainer, ctx);
 
     ctx.containers = {
       root: rootContainer,
