+++ packages/cheeket-koa/lib/index.ts
@@ -1,7 +1,6 @@
-import dependency from "./dependency.middleware";
+import dependency from "./dependency";
 import * as Token from "./token";
 
-export { default as Initializer } from "./initializer";
 export { default as ContainerContext } from "./container-context";
 export { dependency };
 export { Token };
